powerbi_refresher/requirements.txt
Flask==3.0.0
msal==1.24.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
